requirements.txt
flask==2.0.1
python-dotenv==0.19.0
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-wtf==0.15.1
werkzeug==2.0.1
pymysql==1.0.2
cryptography==3.4.7

